refactor main tabs
